Update to version 1.1.0
1. Significant performance improvement!
First, many tasks are now performed in a single cycle instead of multiple ones.
Second, a cache has been added!
The tool now remembers which versions of the models it has already created. Previously, it checked for their existence in the mod's content every time, wasting time. Now it skips that step!
The cache file is stored next to the `.exe` file. If something goes wrong, you can safely delete it, and the program will reanalyze the mod's content to find the ready-made models again.
2. Work has begun on model coloring!
You need to update the FGD file and save every map that already contains `prop_static_scalable` before using the new program version.
3. Fixed a bug where models were extracted from the wrong VPK.
This happened when models with the same names existed in different VPK files.
4. Fixed a bug that caused map compilation to fail if there were no `prop_static_scalable` entities.
5. Improved console log informativeness.
6. Various minor fixes and improvements.